Move Game service to Views folder
